drfbase/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-cleanup==6.0.0
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
Pillow==9.0.1
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
